Tested "getAttachment" method + grouped attachment-related calls
